requirements.txt
plotly
pandas
numpy
math
pillow
nltk
re
requests
streamlit
sklearn
